Find who finished in a certain position of Boston Marathon

The boston_position() function uses the results_boston database to find who finished in a certain position of male and female Boston Marathons from 2009 to 2022.

Usage

boston_position(gender, year, position)

Arguments

gender

Gender of the athlete (string).

year

Year when the event occured (string).

position

Position which the athlete finished the race (integer).

Value

A dataframe of 7 column with year of event, name of the athlete, age of race day, gender, country of citizenship, official time and finishing rank by gender.

Examples

boston_position(gender = "both", year = 2021:2022, position = 1:3)
#> # A tibble: 12 × 7
#>     year full_name      age_on_race_day gender country_of_ctz_name official_time
#>    <int> <chr>                    <dbl> <chr>  <chr>               <Period>     
#>  1  2022 Evans Chebet                33 Male   Kenya               2H 6M 51S    
#>  2  2022 Lawrence Cher…              33 Male   Kenya               2H 7M 21S    
#>  3  2022 Benson Kipruto              31 Male   Kenya               2H 7M 27S    
#>  4  2022 Peres Jepchir…              28 Female Kenya               2H 21M 1S    
#>  5  2022 Ababel Yeshan…              30 Female Ethiopia            2H 21M 5S    
#>  6  2022 Mary Ngugi                  33 Female Kenya               2H 21M 32S   
#>  7  2021 Benson Kipruto              30 Male   Kenya               2H 9M 51S    
#>  8  2021 Lemi Berhanu                27 Male   Ethiopia            2H 10M 37S   
#>  9  2021 Jemal Yimer                 38 Male   Ethiopia            2H 10M 38S   
#> 10  2021 Diana Kipyokei              27 Female Kenya               2H 24M 45S   
#> 11  2021 Edna Kiplagat               41 Female Kenya               2H 25M 9S    
#> 12  2021 Mary Ngugi                  32 Female Kenya               2H 25M 20S   
#> # ℹ 1 more variable: rank_over_gender <dbl>
boston_position(gender = "Female", year = c(2010, 2021:2022), position = 1)
#> # A tibble: 3 × 7
#>    year full_name       age_on_race_day gender country_of_ctz_name official_time
#>   <int> <chr>                     <dbl> <chr>  <chr>               <Period>     
#> 1  2022 Peres Jepchirc…              28 Female Kenya               2H 21M 1S    
#> 2  2021 Diana Kipyokei               27 Female Kenya               2H 24M 45S   
#> 3  2010 Teyba Erkesso                27 Female Ethiopia            2H 26M 11S   
#> # ℹ 1 more variable: rank_over_gender <dbl>
boston_position(gender = "Male", year = 2009, position = 1:10)
#> # A tibble: 10 × 7
#>     year full_name      age_on_race_day gender country_of_ctz_name official_time
#>    <int> <chr>                    <dbl> <chr>  <chr>               <Period>     
#>  1  2009 Deriba Merga                28 Male   Ethiopia            2H 8M 42S    
#>  2  2009 Daniel Rono                 30 Male   Kenya               2H 9M 32S    
#>  3  2009 Ryan Hall                   26 Male   United States of A… 2H 9M 40S    
#>  4  2009 Tekeste Kebede              27 Male   Ethiopia            2H 9M 49S    
#>  5  2009 Robert Cherui…              20 Male   Kenya               2H 10M 6S    
#>  6  2009 Gashaw Asfaw                30 Male   Ethiopia            2H 10M 44S   
#>  7  2009 Solomon Molla               22 Male   Ethiopia            2H 12M 2S    
#>  8  2009 Evans Cheruiy…              26 Male   Kenya               2H 12M 45S   
#>  9  2009 Stephen Kiogo…              34 Male   Kenya               2H 13M 0S    
#> 10  2009 Timothy Cheri…              32 Male   Kenya               2H 13M 4S    
#> # ℹ 1 more variable: rank_over_gender <dbl>